Split Marinara import helpers
